+++ stats/src/sim.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Sim {
-    pub cycles: usize,
-    pub instructions: usize,
+    pub cycles: u64,
+    pub instructions: u64,
 }
